bug(#495): move test to ProgramTest
